Move ServiceRunner to its own module.
